requirements.txt
streamlit==1.28.0
google-generativeai==0.3.1
gspread==5.12.0
google-auth==2.23.4
